# Workaround for bad GPUDirect performance with unaligned GPU buffers